requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.8.2
boto3==1.34.162
botocore==1.34.162
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
Events==0.5
exceptiongroup==1.2.2
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-aws==0.1.18
langchain-community==0.3.20
langchain-core==0.3.49
langchain-google-genai==2.1.2
langchain-text-splitters==0.3.7
langsmith==0.3.22
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.3.1
mypy-extensions==1.0.0
numpy==1.26.4
opensearch-py==2.8.0
orjson==3.10.16
packaging==24.2
pgvector==0.2.5
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==1.26.20
Werkzeug==3.0.6
yarl==1.18.3
zipp==3.20.2
zstandard==0.23.0
